evaluate_main works now
